requirements.txt
discord.py
aiohttp
peewee
flask
requests_oauthlib
requests
BeautifulSoup4
html5lib
peony-twitter